Linting

+++ src/DomainModule/Texture/Rotation.cpp
@@ -36,7 +36,7 @@ Constants::Error Rotation::rotate(int const argc, char** argv) const {
     }
 
     // Create a new texture to hold the rotated result
-    SDL_Texture* rotatedTexture = SDL_CreateTexture(renderer, SDL_PIXELFORMAT_RGBA8888, SDL_TEXTUREACCESS_TARGET, width, height);
+    SDL_Texture* rotatedTexture = SDL_CreateTexture(renderer, SDL_PIXELFORMAT_RGBA8888, SDL_TEXTUREACCESS_TARGET, static_cast<int>(width), static_cast<int>(height));
     if (rotatedTexture == nullptr) {
         return Constants::ErrorTable::TEXTURE::CRITICAL_TEXTURE_MODIFICATION_FAILED();
     }
